python/requirements.txt
# ForgePipeline Python Inference Server
# Install: pip install -r requirements.txt --extra-index-url https://download.pytorch.org/whl/cu124

# PyTorch (CUDA 12.4)
torch==2.10.0
torchvision==0.21.0

# AI model loading
diffusers==0.33.1
transformers==4.49.0
accelerate==1.4.0

# Inference server
fastapi==0.115.8
uvicorn==0.34.0

# Mesh processing
trimesh==4.6.6

# Image handling
Pillow==11.1.0

# Model downloading
huggingface-hub==0.28.1

# VRAM monitoring
pynvml==12.0.0
